Move boot to UEFI

+++ boot/src/main.rs
@@ -1,10 +1,11 @@
 #![no_std]
 #![no_main]
 #![allow(incomplete_features)]
-#![feature(abi_x86_interrupt)]
+#![feature(abi_efiapi)]
 #![feature(decl_macro)]
 #![feature(fn_traits)]
 #![feature(inline_const)]
+#![feature(negative_impls)]
 #![feature(never_type)]
 #![feature(once_cell)]
 #![feature(ptr_metadata)]
@@ -13,8 +14,8 @@
 mod arch;
 
 #[no_mangle]
-fn __lock_disable_sched_save() -> u64 {
+fn __lock_disable_sched_save() -> usize {
     0
 }
 #[no_mangle]
-fn __lock_restore_sched(_: u64) {}
+fn __lock_restore_sched(_: usize) {}
